+++ apps/sv/src/main/scala/com/daml/network/sv/setup/JoiningNodeInitializer.scala
@@ -369,6 +369,7 @@ class JoiningNodeInitializer(
           localDomainNode,
           domainId,
           svcStoreWithIngestion.connection,
+          clock,
           retryProvider,
           logger,
         )