+++ cluster/pulumi/canton-network/src/sv.ts
@@ -288,6 +288,7 @@ export function installSvNode(config: SvConfig): {
           }
         : undefined,
     extraDependsOn: [svApp],
+    useSequencersFromScan: false,
   });
 
   installCNHelmChart(