+++ cluster/pulumi/canton-network/src/validator1.ts
@@ -73,5 +73,6 @@ export async function installValidator1(
     extraDependsOn,
     auth0AppName: 'validator1',
     participantBootstrapDump,
+    useSequencersFromScan: true,
   });
 }